Move datapath into a datapath package
