+++ enforcer/datapath/datapath.go
@@ -16,6 +16,7 @@ import (
 	"github.com/aporeto-inc/trireme-lib/enforcer/acls"
 	"github.com/aporeto-inc/trireme-lib/enforcer/connection"
 	"github.com/aporeto-inc/trireme-lib/enforcer/datapath/nflog"
+	"github.com/aporeto-inc/trireme-lib/enforcer/datapath/proxy/tcp"
 	"github.com/aporeto-inc/trireme-lib/enforcer/utils/fqconfig"
 	"github.com/aporeto-inc/trireme-lib/enforcer/utils/secrets"
 	"github.com/aporeto-inc/trireme-lib/enforcer/utils/tokens"
@@ -152,7 +153,7 @@ func New(
 
 	d.nflogger = nflog.NewNFLogger(11, 10, d.puInfoDelegate, collector)
 	//passing d here since we can reuse the caches and func here rather than redefining them again in proxy.
-	d.proxyhdl = NewProxy(":5000", true, false, d)
+	d.proxyhdl = tcp.NewProxy(":5000", true, false, d)
 	return d
 }
