+++ enforcer/datapath/proxy/tcp/tcp.go
@@ -1,4 +1,4 @@
-package proxy
+package tcp
 
 import (
 	"crypto/tls"
